requirements.txt
amqp==5.1.1
aniso8601==9.0.1
anyio==3.6.2
ariadne==0.16.1
async-timeout==4.0.2
autopep8==2.0.0
billiard==3.6.4.0
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
dnspython==2.2.1
fastapi==0.85.0
graphql-core==3.2.3
graphql-relay==3.2.0
h11==0.14.0
httptools==0.5.0
idna==3.4
kombu==5.2.4
packaging==21.3
prompt-toolkit==3.0.33
pycodestyle==2.10.0
pydantic==1.10.2
pymongo==4.3.3
pyparsing==3.0.9
python-dotenv==0.21.0
pytz==2022.6
PyYAML==6.0
redis==4.3.5
six==1.16.0
sniffio==1.3.0
starlette==0.20.4
tomli==2.0.1
typing-extensions==4.4.0
uvicorn==0.20.0
uvloop==0.17.0
vine==5.0.0
watchfiles==0.18.1
wcwidth==0.2.5
websockets==10.4
